requirements.txt
pygame>=2.5.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
ollama>=0.1.0
matplotlib>=3.7.0
dataclasses-json>=0.6.0
PyYAML>=6.0
openai>=1.0.0 